requirements-dev.txt
matplotlib
jupyter
scikit-learn
pytest
pytest-cov
coverage
flake8
pylint
sphinx
sphinx_rtd_theme
nbsphinx
numpydoc
twine
